+++ app/src/provider/SearchProvider.tsx
@@ -35,11 +35,17 @@ const SearchContext = React.createContext<SearchContextType>(
   {} as SearchContextType,
 );
 
+export const LOCALSTORAGE_QUALITY_FILTER = "tidarr-quality-filter";
+
 export function SearchProvider({ children }: { children: ReactNode }) {
   const [loading, setLoading] = useState<boolean>(false);
   const [page, setPage] = useState<number>(1);
   const [keywords, setKeywords] = useState<string>();
-  const [quality, setQuality] = useState<QualityType>("all");
+  const [quality, setQuality] = useState<QualityType>(
+    (window._env_.REACT_APP_TIDARR_DEFAULT_QUALITY_FILTER as QualityType) ||
+      (localStorage.getItem(LOCALSTORAGE_QUALITY_FILTER) as QualityType) ||
+      "all",
+  );
   const [artistPagerLoading, setArtistPagerLoading] = useState<number>();
   const [searchResults, setSearchResults] = useState<TidalResponseType>(
     {} as TidalResponseType,